+++ Espoir/chapter3/chapter3/src/main/java/team/arryn/aa/chapter3/service/UserService.java
@@ -6,7 +6,7 @@ import team.arryn.aa.chapter3.util.Result;
 
 public interface UserService {
     Result addUser(User user);
-    Result getUserById(int id);
+    Result getUserById(Long id);
     Result updateUser(User user);
-    Result delUser(int id);
+    Result delUser(Long id);
 }
